+++ docs/HISTORY.md
@@ -3,3 +3,7 @@
 - Client: source code of the client app
 - Server: source code of the server api
 
+
+# Server
+
+- Use redis to cache